+++ llm_attacks_bishopfox/util/util_functions.py
@@ -509,7 +509,8 @@ def find_last_occurrence_of_array_in_array(inner_array, outer_array, start_index
 
 # Convert a nice array of command elements to a terrible single value which can be used with bash -c.
 # Because it's 2024, and Python still makes it very hard to capture stderr + stdout exactly the way that it would appear in a shell, without deadlocking or running out of buffer space, and while letting the developer set a timeout on execution without some kind of hokey second thread
-def command_array_to_bash_c_argument(command_array):
+# Or, alternatively, log a reproducible (properly quoted) string that represents the command that launched a Python script, so the user can re-run it later
+def command_array_to_string(command_array):
     inner_command = None
     for i in range(0, len(command_array)):
         current_element = shlex.quote(command_array[i])
@@ -519,9 +520,8 @@ def command_array_to_bash_c_argument(command_array):
             inner_command = f"{inner_command} {current_element}"
     #result = shlex.quote(inner_command)
     result = inner_command
-    #print(f"[command_array_to_bash_c_argument] Debug: input = {command_array}, output = {result}")
+    #print(f"[command_array_to_string] Debug: input = {command_array}, output = {result}")
     return result
-
 
 # return a slice that begins with the lower of two other slices start values, and ends with the greater of their stop values.
 # Not quite a "union" operation, so I've avoided that term.